Simplify js_library, remove "__internal__node_modules"
This changes the output of ts_library. Instead of having the srcs at the root of the generated directory, they're in a nested path equivalent to their placement in the workspace.

This means that directory structure is always kept, and things like relative CSS import statements can't be broken anymore.