Ran buildifier to fix usage of single_file attribute (#57)
Fixes #56

This change is the result of running buildifier --lint=fix on all Starlark and BUILD files, which automatically changed all usages of the single_file and allow_files attributes to allow_single_file. This avoids the need to pass --incompatible_disable_deprecated_attr_params when Bazel 0.27 becomes available.

It also cleaned up a few load statements as a side effect.